Changes Form names. Formats AutoCallDisplay form. Resolves #10. Creates justCalled string var to save called number into AutoCallDisplay
